Server side: +configuration auth for customer done +modifying controllers for cart routes +RESTAPI Doc done

+++ server/controllers/CartController.js
@@ -1,4 +1,5 @@
 const {Cart, Product} = require(`../models`)
+const e = require("express")
 
 class CartController {
 
@@ -10,7 +11,7 @@ class CartController {
             message: `Sorry can't find data.`
         }
 
-        Cart.findAll({where: {UserId}, include: [`Product`, `User`], order: [['createdAt', 'ASC']]})
+        Cart.findAll({where: {UserId}, include: [`Product`], order: [['createdAt', 'ASC']]})
         .then((result) => {
             if (result) {
                 res.status(200).json(result)
@@ -26,13 +27,24 @@ class CartController {
     static add(req, res, next) {
         let newCart = {
             UserId: req.user.id,
-            ProductId: req.params.id,
+            ProductId: req.params.ProductId,
             quantity: req.body.quantity,
             status: false
         }
-        Cart.create(newCart)
+        Cart.findOne({where: {UserId: newCart.UserId, ProductId: newCart.ProductId}})
         .then((result) => {
-            res.status(201).json(result)
+            if(!result) {
+                return Cart.create(newCart)
+            } else {
+                return Cart.increment({quantity: newCart.quantity}, {where: {id: result.id}})
+            }
+        })
+        .then((result) => {
+            if(Array.isArray(result)) {
+                res.status(200).json(result[0][0][0])
+            } else {
+                res.status(201).json(result)
+            }
         })
         .catch((err) => {
             next(err)
@@ -40,15 +52,14 @@ class CartController {
     }
 
     static delete(req, res, next) {
-        let ProductId = req.params.id
-        let UserId = req.user.id
+        let id = req.params.id
         const error = {
             name: `otherError`,
             statusCode: 404,
             message: `Sorry can't find data.`
         }
 
-        Cart.destroy({where: {UserId, ProductId}})
+        Cart.destroy({where: {id}})
         .then((result) => {
             if (result === 0) {
                 throw error
@@ -62,17 +73,19 @@ class CartController {
     }
 
     static deleteAll(req, res, next) {
+        let UserId = req.user.id
         const error = {
             name: `otherError`,
             statusCode: 404,
             message: `Sorry can't find data.`
         }
-        Cart.destroy({where: {}})
+
+        Cart.destroy({where: {UserId}})
         .then((result) => {
             if (result === 0) {
                 throw error
             } else {
-                res.status(200).json({message: `Successfully delete all items from cart!`})
+                res.status(200).json({message: `Successfully delete items from cart!`})
             }
         })
         .catch((err) => {
